requirements.txt
# 温度控制系统 - 整体项目依赖文件
# Temperature Control System - Complete Project Dependencies
# 适用于整个系统：EdgeX + eKuiper + EMQX + MCP

# ========================================
# 使用说明
# ========================================
# 1. 创建虚拟环境:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    # venv\Scripts\activate   # Windows
#
# 2. 安装依赖:
#    pip install -r requirements.txt
#
# 3. 系统组件:
#    - EdgeX Foundry: Docker 容器部署
#    - eKuiper: Docker 容器部署  
#    - EMQX Cloud: 云服务
#    - MCP Server: Python 应用

# ========================================
# 核心 MCP 和 MQTT 依赖
# ========================================
mcp>=1.0.0                    # Model Context Protocol 服务器框架
fastmcp>=2.5.0                # Fast MCP 服务器实现
paho-mqtt>=2.1.0              # MQTT 客户端 (与 EMQX 通信)
pydantic>=2.0.0               # 数据验证和设置管理
python-dotenv>=1.0.0          # 环境变量管理
requests>=2.31.0              # HTTP 客户端 (eKuiper API, EMQX API)
anyio>=4.0.0                  # 异步 I/O 支持
httpx>=0.25.0                 # 现代异步 HTTP 客户端

# ========================================
# 数据处理和验证
# ========================================
jsonschema>=4.0.0             # JSON 模式验证

# ========================================
# 系统监控和工具
# ========================================
psutil>=5.9.0                 # 系统资源监控
colorama>=0.4.6               # 终端彩色输出

# ========================================
# 安全和加密
# ========================================
cryptography>=41.0.0          # SSL/TLS 加密支持

# ========================================
# 开发和测试工具
# ========================================
pytest>=7.0.0                 # 测试框架
pytest-asyncio>=0.23.0        # 异步测试支持
black>=23.0.0                 # 代码格式化工具
isort>=5.12.0                 # 导入语句排序
flake8>=6.0.0                 # 代码风格检查
mypy>=1.0.0                   # 静态类型检查

# ========================================
# 可选的数据分析依赖 (如需要)
# ========================================
# pandas>=2.0.0               # 数据分析和处理 (取消注释以启用)
# numpy>=1.24.0               # 数值计算 (取消注释以启用)
# matplotlib>=3.7.0           # 数据可视化 (取消注释以启用)

# ========================================
# 可选的数据分析依赖 (如需要)
# ========================================
# pandas>=2.0.0               # 数据分析和处理
# numpy>=1.24.0               # 数值计算
# matplotlib>=3.7.0           # 数据可视化

# ========================================
# 系统架构说明
# ========================================
#
# 本项目采用以下技术栈:
#
# 📱 Claude Desktop (MCP 客户端)
#     ↕️ MCP Protocol
# 🔌 EMQX MCP Server (本文件依赖)
#     ↕️ MQTT over SSL/TLS  
# ☁️ EMQX Cloud (托管 MQTT Broker)
#     ↕️ MQTT Topics
# ⚡ eKuiper (Docker 容器)
#     ↕️ REST API
# 🏭 EdgeX Foundry (Docker 容器)
#     ↕️ Virtual Device Protocol
# 🌡️ 虚拟温湿度传感器
#
# Docker 组件无需 Python 依赖，通过 REST API 交互。
# 本 requirements.txt 主要用于 MCP Server 和监控脚本。

# ========================================
# 安装示例
# ========================================
#
# 最小安装 (仅核心功能):
# pip install mcp fastmcp paho-mqtt requests python-dotenv
#
# 开发环境 (包含测试工具):
# pip install -r requirements.txt
#
# 生产环境 (排除开发工具):
# pip install mcp fastmcp paho-mqtt pydantic python-dotenv requests httpx psutil colorama cryptography
